+++ package.json
@@ -13,10 +13,10 @@
     "start:rest": "nodemon -L",
     "build": "npm run clean && npm run compile",
     "cli:help": "npm run ts ./src/main.cli.ts -- --help",
-    "cli:version": "npm run ts ./src/main.cli.ts --version",
-    "cli:import": "npm run ts ./src/main.cli.ts --import mocks/mock-data.generated.tsv mongodb://admin:admin@127.0.0.1:27017/six-cities?authSource=admin salt",
-    "cli:import:docker": "npm run ts ./src/main.cli.ts --import mocks/mock-data.generated.tsv mongodb://admin:admin@db:27017/six-cities?authSource=admin salt",
-    "cli:generate": "npm run ts ./src/main.cli.ts --generate 10 mocks/mock-data.generated.tsv http://localhost:3000/api",
+    "cli:version": "npm run ts ./src/main.cli.ts -- --version",
+    "cli:import": "npm run ts ./src/main.cli.ts -- --import mocks/mock-data.generated.tsv mongodb://admin:admin@127.0.0.1:27017/six-cities?authSource=admin salt",
+    "cli:import:docker": "npm run ts ./src/main.cli.ts -- --import mocks/mock-data.generated.tsv mongodb://admin:admin@db:27017/six-cities?authSource=admin salt",
+    "cli:generate": "npm run ts ./src/main.cli.ts -- --generate 10 mocks/mock-data.generated.tsv http://localhost:3000/api",
     "docker": "docker compose --file ./docker-compose.yml --env-file ./.env --project-name \"six-cities\" up -d",
     "mock:server": "json-server ./mocks/mock-server-data.json --port 3000",
     "lint": "eslint ./src/ --ext .ts",
