--- src/shared/modules/auth/auth.constant.ts
@@ -1,2 +1,0 @@
-export const JWT_ALGORITHM = 'HS256';
-export const JWT_EXPIRED = '2d';
