+++ src/shared/modules/auth/errors/base-user.exception.ts
@@ -1,4 +1,4 @@
-import { HttpError } from '../../../libs/rest/index.js';
+import { HttpError } from '../../../libs/index.js';
 
 export class BaseUserException extends HttpError {
   constructor(httpStatusCode: number, message: string) {
