+++ src/shared/modules/category/index.ts
@@ -1,5 +1,5 @@
-export { CategoryEntity, CategoryModel } from './category.entity.js';
-export { CreateCategoryDto } from './dto/create-category.dto.js';
-export { CategoryService } from './category-service.interface.js';
 export { DefaultCategoryService } from './default-category.service.js';
+export { CategoryEntity, CategoryModel } from './category.entity.js';
+export { CategoryService } from './category-service.interface.js';
 export { createCategoryContainer } from './category.container.js';
+export { CreateCategoryDto } from './dto/create-category.dto.js';
