--- src/shared/modules/comment/comment.constant.ts
@@ -1,1 +1,0 @@
-export const DEFAULT_COMMENT_COUNT = 50;
