+++ src/shared/modules/comment/comment.entity.ts
@@ -1,17 +1,10 @@
 import { defaultClasses, getModelForClass, modelOptions, prop, Ref } from '@typegoose/typegoose';
 import { UserEntity } from '../user/index.js';
-import { OfferEntity } from '../offer/offer.entity.js';
+import { OfferEntity } from '../offer/index.js';
 
-// eslint-disable-next-line @typescript-eslint/no-unsafe-declaration-merging
 export interface CommentEntity extends defaultClasses.Base {}
 
-@modelOptions({
-  schemaOptions: {
-    collection: 'comments',
-    timestamps: true,
-  },
-})
-// eslint-disable-next-line @typescript-eslint/no-unsafe-declaration-merging
+@modelOptions({ schemaOptions: { collection: 'comments', timestamps: true } })
 export class CommentEntity extends defaultClasses.TimeStamps {
   @prop({ trim: true, required: true })
   public text: string;
