+++ src/shared/modules/offer/index.ts
@@ -1,7 +1,7 @@
-export { OfferEntity, OfferModel } from './offer.entity.js';
-export { CreateOfferDto } from './dto/create-offer.dto.js';
-export { OfferService } from './offer-service.interface.js';
-export { DefaultOfferService } from './default-offer.service.js';
-export { createOfferContainer } from './offer.container.js';
-export { OfferRdo } from './rdo/offer.rdo.js';
 export { CreateOfferValidationMessage } from './dto/create-offer.messages.js';
+export { DefaultOfferService } from './default-offer.service.js';
+export { OfferEntity, OfferModel } from './offer.entity.js';
+export { OfferService } from './offer-service.interface.js';
+export { createOfferContainer } from './offer.container.js';
+export { CreateOfferDto } from './dto/create-offer.dto.js';
+export { OfferRdo } from './rdo/offer.rdo.js';
