+++ src/shared/modules/offer/offer.container.ts
@@ -1,5 +1,6 @@
 import { types } from '@typegoose/typegoose';
 import { Container } from 'inversify';
+
 import { DefaultOfferService } from './default-offer.service.js';
 import { OfferEntity, OfferModel } from './offer.entity.js';
 import { OfferService } from './offer-service.interface.js';
