+++ src/shared/modules/offer/type/param-offerid.type.ts
@@ -1,7 +1,3 @@
 import { ParamsDictionary } from 'express-serve-static-core';
 
-export type ParamOfferId =
-  | {
-      offerId: string;
-    }
-  | ParamsDictionary;
+export type ParamOfferId = { offerId: string } | ParamsDictionary;
