+++ src/shared/modules/user/default-user.service.ts
@@ -1,11 +1,12 @@
-import { DocumentType, types } from '@typegoose/typegoose';
-import { CreateUserDto } from './dto/create-user.dto.js';
-import { UserService } from './user-service.interface.js';
-import { UserEntity } from './user.entity.js';
 import { inject, injectable } from 'inversify';
-import { Component } from '../../types/index.js';
-import { Logger } from '../../libs/logger/index.js';
+import { DocumentType, types } from '@typegoose/typegoose';
+
+import { UserService } from './user-service.interface.js';
+import { CreateUserDto } from './dto/create-user.dto.js';
 import { UpdateUserDto } from './dto/update-user.dto.js';
+import { Component } from '../../types/index.js';
+import { UserEntity } from './user.entity.js';
+import { Logger } from '../../libs/index.js';
 
 @injectable()
 export class DefaultUserService implements UserService {
