+++ src/shared/modules/user/dto/login-user.dto.ts
@@ -1,6 +1,6 @@
 import { IsEmail, IsString, Length } from 'class-validator';
-import { CreateLoginUserMessage } from './login-user.messages.js';
 import { Expose } from 'class-transformer';
+import { CreateLoginUserMessage } from './login-user.messages.js';
 
 export class LoginUserDto {
   @IsEmail({}, { message: CreateLoginUserMessage.email.invalidFormat })
