+++ src/shared/modules/user/index.ts
@@ -1,8 +1,8 @@
+export { DefaultUserService } from './default-user.service.js';
+export { UserService } from './user-service.interface.js';
+export { createUserContainer } from './user.container.js';
 export { UserEntity, UserModel } from './user.entity.js';
 export { CreateUserDto } from './dto/create-user.dto.js';
-export { UserService } from './user-service.interface.js';
-export { DefaultUserService } from './default-user.service.js';
-export { createUserContainer } from './user.container.js';
+export { LoginUserDto } from './dto/login-user.dto.js';
 export { UserController } from './user.controller.js';
 export { UserRdo } from './rdo/user.rdo.js';
-export { LoginUserDto } from './dto/login-user.dto.js';
