+++ src/shared/modules/user/type/create-user-request.type.ts
@@ -1,6 +1,6 @@
 import { Request } from 'express';
 
-import { RequestBody, RequestParams } from '../../../libs/rest/index.js';
+import { RequestBody, RequestParams } from '../../../libs/index.js';
 import { CreateUserDto } from '../dto/create-user.dto.js';
 
 export type CreateUserRequest = Request<RequestParams, RequestBody, CreateUserDto>;
