+++ src/shared/modules/user/type/login-user-request.type.ts
@@ -1,6 +1,6 @@
 import { Request } from 'express';
 
-import { RequestBody, RequestParams } from '../../../libs/rest/index.js';
+import { RequestBody, RequestParams } from '../../../libs/index.js';
 import { LoginUserDto } from '../dto/login-user.dto.js';
 
 export type LoginUserRequest = Request<RequestParams, RequestBody, LoginUserDto>;
