+++ src/shared/modules/user/user.container.ts
@@ -1,11 +1,12 @@
 import { Container } from 'inversify';
-import { UserService } from './user-service.interface.js';
-import { Component } from '../../types/index.js';
-import { DefaultUserService } from './default-user.service.js';
 import { types } from '@typegoose/typegoose';
+
+import { DefaultUserService } from './default-user.service.js';
+import { UserService } from './user-service.interface.js';
 import { UserEntity, UserModel } from './user.entity.js';
-import { Controller } from '../../libs/index.js';
 import { UserController } from './user.controller.js';
+import { Component } from '../../types/index.js';
+import { Controller } from '../../libs/index.js';
 
 export function createUserContainer() {
   const userContainer = new Container();
