+++ src/shared/types/index.ts
@@ -1,9 +1,10 @@
 export { DocumentExists } from './document-exists.interface.js';
 
 export { OfferType, OfferComfort, City, Offer } from './offer.type.js';
-export { ChalkColorMethod, Comment } from './comment.type.js';
+export { ChalkColorMethod } from './common.type.js';
 export { User, UserType } from './user.type.js';
 export { Category } from './category.type.js';
+export { Comment } from './comment.type.js';
 
 export { Component } from './component.enum.js';
 export { SortType } from './sort-type.enum.js';
